Add support of submodules to init_module_args()
This function is used to init the argument parser from modules.
The function goes through installed modules and does the init only for
the requested modules.
But if the expected function 'init_args' is not in the module but in a
submodule then it skips it, which is not the expected behavior.
Update the function to support also submodules.

Signed-off-by: Alexandre Bailon <alexandre.bailon@gmail.com>