Remove libregice args from regicecommon
The args of libregice were defined and processed in regicecommon.
This was required because it was required by any application using
libregice. But, in meantime, this was not a good place.
This updates and completes regicecommon to allow to define and process
arguments in the module where they belong.

Basically, this allows to define clients and svd arguments in libregice,
and to allocate the client, the svd and the device in libregice.

Signed-off-by: Alexandre Bailon <alexandre.bailon@gmail.com>